+++ packages/template-sveltekit/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @happyvertical/smrt-template-sveltekit
 
+## 0.29.26
+
 ## 0.29.25
 
 ## 0.29.24